Errors in the Package config and the Webpack Config
path.resolve() resolves two paths into an absolute path if one is an absolute
path and the other is a relative path. The problem with the Webpack config was
that I was tryng to resolve __dirname with /app/index.html -- I was also making
this mistake with the output property -- and the result was /app/index.html.
Webpack was looking at the root for a folder called app and failing to find
index.html. The problem is that I was giving resolve two absolute paths for it
to resolve; resolve() when given two absolute path resolves to the latter path
because it sees them two distinct paths. So what should have been written was
path.resolve(__dirname, './app/index.html'). Now you have an absolute path and a
relative path that resolves to the same value as __dirname+'/app/index.html'.

The behavior of path is that it tries to simulate changing directory in an
abstract manner. It has no knowledge of the current directory. So it will take
the first path and try to move to the second path, provided that it's a relative
path. If the two paths are absolute paths with no subpath in common, then it
will just return the second path. If they do have something in common, it will
return the largest absolute path possible. Essentially what resolve does is it
tries to return a NORMALIZED absolute path -- that is, an absolute path that can
be formed from a series of paths.

webpack.config.js:
        - The error was giving resolve() two absolute paths, when it should have
          been given an absolute path and a relative path.

        - Fixed the path.resolve issues by changing the absolute path
          '/app/index.html' to a relative path './app/index.html'
                - Similarly for the path in output
package.json:
        - I forgot to include babel-preset-env. This module is important since
          it allows webpack to translate ES6,7 into the current ES standard.
          Without it we couldn't use an new, unsupported js features.