Adding the App and Popular Components
App contains the components and elements that make up the app. Currently, it
imports the Popular component which consists of a nav bar created using an
unordered list.

The Popular component currently contains a list of languages styled as a
navigation bar. In the future, we'll incorporate the ability to hold and change
the state of the bar, and we'll add the components that fetch and display data
based on what language has been selected.

index.css:
        - This css file affects the entire app. The styling for Popular is found
          here.

        - The unordered list, languages, in Popular uses flexbox.

App.js:
        - Contains the app

        - Imports the Popular component as a part of the app

Popular.js:
        - Currently, it contains a unordered list of languages styled as a
          navigation bar.