Adding State to the Popular Component
The Popular component should be a clickable component that remembers its state,
that is, the current selection.

State is added to the component by using the state property of an instance
accessible by this.state. The initial state is initialized in the constructor of
the class so that when the class is instantiated (the creation of a instance
occurs) the state is immediately initialized.

The state changes for this component, we update the state for an instance
using the instance's setState method. The setState method takes a callback that
returns the desired state. The reason setState takes a callback is that there
is a process that must occur if the state is changed -- that is, the React Life
Cycle. The state is set during the process, and not immediately, thus the need
to invoke a callback. How and when the state is set is determined by the React
Life Cycle. It's important to know what this process is as it becomes incredibly
useful when making API calls to populate the component with data.

"This" becomes a tricky part of implementing state. First, there is the problem
in which we want to access the state in map's callback. We want to be able to
write "this.state" to get the state of the component. Unfortunately, the function
on invocation is bound to the global object, therefore, "this" refers to the
global object. This is just how javascript operates. How do we redefine the
meaning of "this"? We have to explicitly bind it to the instance. Luckily, map
allows us to redefine "this" for the callback; the object that the callback will
be bound to can be passed as the second argument. So, we can pass "this", the
instance reference, with the callback and expect any references to "this" in the
callback to refer to the instance.

We also want to update the app on a click event; this requires us to present a
handler that provides a procedure to do this. So we create a method
updateLanguage to do update the selected language of the state. For click
events, we also set up a listener onClick for each li with updateLanguage as the
handler -- updateLanguage will be invoked on a click. Now we are presented with
a problem: updateLanguage is no longer used as a method and, as a result, "this"
of this.setState no longer refers to the instance. By setting updateLanguage as
a handler, it's going to be invoked differently and not as a method of an
instance of Popular. So "this" will no longer refer to the instance (instead,
it'll refer to null, it should be the window object. Why?). How can we update
the state away from the instance? We bind updateLanguage to an instance and we
do it immediately in the constructor. Once it is bound to the instance, no
matter where function is passed to and invoked, "this" will always refer to the
instance.

Now we are faced with another problem. If we simply assign updateLanguage as
such: <li onClick=this.updateLanguage>, there is no way we can control what
arguments get passed. We want to be able to pass the language. There are
different ways to do this. We could get the value of the DOM element and update
the state in updateLanguage. Instead, we'll bind updateLanguage's argument to
the current value of lang. This way when it gets invoked later, it'll use that
value to update the language. Also note that we skip binding updateLanguage to
an object; We've already bounded it an instance of Popular, so, there is no
reason override that bind.

Yup, there's a lot to do when implementing state. But, once it's set up, the state
of a component is a lot easier to manage.

On another note, don't be fooled by the ES6 sugar. The methods defined in the
class are properties of the object (not the class). In fact, you can now do this
in ES6:
var obj = {
        yo() {
        /*code*/
        }
        boo() {
        /*code*/
        }
}
That is the shorthand of this:
var obj = {
        yo: function () {
        /*code*/
        boo: function () {
        /*code*/
        }
}

Popular.js:
        - Adding state to the component
                - The component now remembers the state of its selected
                  language. In addition, it uses the state to apply the correct
                  styling to element that corresponds to the selected language.

                - The initial state is initialized in the constructor. It sets
                  the selected language to all. The updateLanguage function is
                  immediately bounded to the instance because it is used as a
                  click handler to update the language on clicks.

                        - Subtle mistakes to make are:
                                - onClick=this.updateLanguage(lang)
                                        - This assigns the result, and not the
                                          function like we want, as the handler.
                                        - Should be
                                          onClick=this.updateLanguage.bind(null,lang)

                                - Forgetting to bind updateLanguage to the
                                  instance and then passing it onto onClick.

                                - Forgetting to bind updateLanguage's argument
                                  to lang