Moving the Language Nav Bar to it's own component
The idea is for the Popular component to encapsulate the language bar and a grid
of popular repositories associated with that language. If we continue to add
functionality to this component, it's going to add to its intricacy. Therefore,
to alleviate some of the mental overhead, the render code is moved into its own
component called LanguageBar. In doing so, we've separated the stateless part of
the component from the stateful part.

It's important to notice that we are placing the state high above the stateless
components in a containing component, The reason for this is that as we
manipulate the state, the changes can be passed down as props to the stateless
components when a re-render occurs (setState causes a re-render). Also, it
allows the state to be centralized, making it easier to manipulate and keep
track of it. This pattern is called the Container pattern, and, in our case,
Popular is the container component.

In moving the render code of Popular into its own component, we have to make
make some changes. First, any references to "this" have to be examined, they no
longer may refer. Second, we don't have access to updateLanguage anymore since
we are in a different component. Third, we can't access the state, as previously
stated, because we are in a different component. We have to address these
issues.

Since the state is contained in Popular and LanguageBar is rendered in Popular,
we can pass the state as a prop and LanguageBar can access them through
this.props.  Similarly, we can do this with updateLanguage. We can pass
updateLanguage as a prop, and, since it's already bound to an instance of
Popular, we can be certain that it will update Popular's state.

Finally, we render LanguageBar in the Popular component, passing it down state
and updateLanugage as props.

As a side note, components are usually defined in their own file as a module. This
allows us to import them as we need them. However, since the language nav bar
is specific to the Popular component, we keep it in the same file as Popular. If
in the future we need to use it elsewhere, we'll move LanguageBar into its own
file.

Popular.js:
        - Moving the render code into its own stateless component, LanguageBar.
                - In doing so, Popular becomes a container component housing the
                  state away from the stateless components. State changes are
                  filtered down as props.

        - The Popular container render the LanguageBar component.