Adding PropTypes for Popular
PropTypes are great way to enforce the types that props of a component will
accept. Also, its great for documenting what props a component requires. Someone
can easily glance at the PropTypes of a component and know how it should be
used.

PropTypes is a property of a Component's class and it takes in an object of
props as the properties and their PropTypes as values. Use isRequired to enforce
the use of that prop. For example:

Yoko.PropTypes {
        ono: PropTypes.string.isRequired,
        yo: PropTypes.number
}

Types are lower case properties of PropTypes. For function and boolean, the
prop types are written as func and bool respectively. Function and boolean are
reserved keywords in js, therefore they can't be used as prop types. For
example, doing named imports from PropTypes will cause a conflict.

import {string, number, boolean, function, arrayOf} from 'prop-types'

Note: We can use the shape method for objects if we need to enforce the structure
of a prop object. For example:

Yoko.PropTypes = {
        list: PropTypes.arrayOf(PropTypes.shape({
                name: PropTypes.string.isRequired,
                date: PropTypes.string
        })
}

Popular.js:
        - Adding PropTypes to LanguageBar class

package.json:
        - Adding prop-types as a dependency