Fetching GitHub Data and Building the Repository Component
We want to be able to populate the app with data from GitHub. In order to do so, we
implement a module to fetch the data, as well as the component to
house the data. We also integrate the fetching mechanism into the
Popular container so that the state can be updated with the data. This process is
done in three phases and we begin with describing the module dedicated to fetch
data from GitHub.

The API helper module is a collection of helper functions that fetch data from
GitHub based on the parameters they're given. The module makes use of the axios
module, a Promise-based HTTP client capable of making AJAX requests, among other
things. The Promise functionality primarily allows us to avoid the problem that
plagues us when programming asynchronously, multiple levels of callbacks, or otherwise
known as callback hell. Promises allow us to chain together callbacks in a
sequence, promising to pass the results as an argument to the next callback in
the chain. Additionally, Promises also promise to capture an event once it
occurs, something that can't be guaranteed by event listeners. The module has
only one method so far, getPopularRepos; it makes an AJAX request to GitHub, using a
URI with the necessary query/filter parameters, for all the popular repositories
written in a particular language. The language depends on the selection made by
the user.  The method returns a promise, the .then method returns a promise,
allowing us to append another callback, one that updates the state of the
component, to the chain.

The Popular container component is tasked with containing and maintaining its
state. As of now, its state consists of the language selected. Now, we need it to
also house the GiitHub data based on the language selected. In other words, on
the selection of a language, it will have to update the state of the data. We
begin by adding an initial state. Previously, we would have had to use the React
life cycle method called getInitialState when using React.createClass(). However,
with the prevalence of ES6 (ES2015) classes, React allows us to use the
constructor to set the initial state. We add the field "repos" to the initial
state and set it to null. Since we need to update the repos when the language
updates, we can fetch the data in the method updateLanguage. We import our
APIHelper module, make the call to getPopularRepos, and attach a callback that
updates the state with the repos when they have been retrieved.

Next, we create the UI that will display the repo data. Since Popular maintains
the state, then, naturally, we create a stateless component and have Popular
pass the component the repos as props. The component is a list of repositories,
who themselves are a list of data. We map the data of each repo to a heirarchy
of list elements. The lists are then styled, in the stylesheet, into a grid
using Flexbox. One note of caution, don't set mindlessly set the key to all li's
as the index. You'll just get multiple li's with the same key, which makes no
sense since the key is suppose to be unique -- Yeah, I was somewhere else when I
did that.

Before we integrate the RepoGrid into the Popular component. We can render
the data as text to get a better glimpse of how the page loads. You'll notice
that when we initially load the page, the page will default to the language 'All'
and a null. That's a bad look, but it's to be expected because that's our
initial state. So to fix this, we make use of the React life cycle method
componentDidMount.  This method allows us to fetch data and update the state,
after the component has been mounted on the DOM. (Now would be a good time to
ask why do we fetch data in this method and not in componentWillMount? I don't
know the answer just yet) In componentDidMount, we make a call to updateLanguage
to update the state of the Popular Component (which includes fetching data from
GitHub). Setting the state will cause a re-render of the component. On the
initial load of the page, the page should still show the initial state of null
repos with 'All' selected.  However, it should also automatically re-render and
populate the page with Repos. We'll take care of the null issue next.

As we place the RepoGrid component in the render method of Popular, we have to
keep in mind that RepoGrid expects the data to be passed down to it so that it
can render it. The problem with fetching data is that it takes time and, most of
the time, it won't be available by the time Popular invokes its render method.
So it passes null down to the RepoGrid's repos prop. When RepoGrid goes to
render, it wont' be able to as .map doesn't exists for null, causing an error.
If you declare PropTypes for the component, you'll get a PropTypes error
instead. Therefore, we have Popular's render method display a loading message if
the repos field of the state is null, otherwise, render the RepoGrid with the
non-null results.  This should prevent RepoGrid from rendering with a null repos
prop. Additionally, it also takes care of that null issue when initially
rendering the page; the "loading" message is displayed instead of null.

We also reset the repos to null every time a new language is selected. This
causes the "loading" message appear every time a new language is selected -- not
resetting the repos field causes the current repos to linger for a short time on
the page while the new data is being fetched. It give the impression of lag or
non-responsiveness. That's why we reset the repos to null which causes the
loading message to be displayed. It's better we explicitly say that we are
loading the results rather than give the impression that the page is not
responding.

Popular.js:
        - Add the repos field to the state

        - UpdateLanguage fetches repos from GitHub and updates the state when it
          recieves the data.

        - RepoGrid is the stateless component that will display the fetched
          Repos.

        - To make sure that RepoGrid doesn't take on null props from the initial
          state, we check if the repos is null before trying to render an
          instance of RepoGrid. If the repos field in the state is null, then we
          display a "Loading message instead.

        - To remedy the initial load of an empty page, we run updateLanguage in
          componentDidMount, a lifecycle method that runs after the component
          has been mounted to the DOM, to fetch the Popular Repos for all
          languages and update the state, causing an automatic re-render of the
          component.

        - Which PropTypes are accepted or required by RepoGrid is also declared
          in this file.

index.css:
        - Add styling to elements in the RepoGrid. Flexbox is used to create the
          grid.

package.json:
        - Set axios as a dependency

/utils/APIHelpers.json:
        - Adding the getPopularRepos method to the module. It uses axios to
          fetch the Popular Repos from GitHub. Axios is a Promise-based HTTP
          client.