Adding React-Router, a Navigation Bar, and New Pages/Sections
We want to be able to click on a link and be directed to the component
associated with the path.  Vanilla <a> tags, unfortunately, won't work because
their default behavior is to send requests to the server for the page that the
link is associated with. The webpack server isn't set up for routing, so we'll
add the react-router module to help us with the routing.

react-router is a module that allows us to associate paths to components.  It
uses the component metaphor to abstract away the routing functionality. Since
paths are being associated with components, the app is no longer rendered
directly using ReactDOM.render(). Instead, the router component will encapsulate
all the routes, directly or indirectly, and we'll render the router to render
the app.

Note: It would be best to follow the react-router tutorial before continuing. (I
find the old one to be we'll written; the new tutorial is weak -- it looks like
they created an app to exhibit examples rather than walk you through it.) I can
only explain what's used in this app in the most concise way. UPDATE: the old
tutorial, while a good tutorial, for the most part, is outdated.  Turns out the
new tutorial isn't a tutorial at all, but documentation -- I don't know how I
missed that. You can still go through the old tutorial and amend what you've
learned.

First, we create the navigation bar to set up links to different paths, paths
that are associated to components. Clicking a link will cause the router to
render the component associated with the link's path. The navigation bar is
relatively simple. It's an unordered list with three links to /home, /battle,
and /popular. The styling uses flexbox to get them in a line. The links, as
previously discussed, aren't regular links. Nav Links are components that add
functionality to <a> such as keeping track of when a path is active, active
styles, and deferring to the router to render the next component. I'd imagine
they prevent the default behavior(.preventDefault()) of <a> and tell the router
to render the component associated with the link's path. Anyway, we declare the
router and it's routes.

The Router is contained in the App component. Router contains a div that
contains the navigation bar, within it are three routes. We need the navigation
bar to always display, therefore, it isn't associated to a route. Instead, it's
rendered as a regular component. (Note: In the past, we would have to associate
it the root path / as an index route.) First, we add the Popular route to the
router, then, the /popular path is associated to the Popular component. Clicking
on the Popular link in the navigation bar will cause the Popular component to be
active. Next, we'll create the Home and Battle components and add the routes.

A few things you have to understand about react-router. Previously, all routes
would be declared and contained within the Router component. Nested routes would
help you to render components as part of a Parent component by assigning them to
props as children, which would allow you to render them by referencing
{this.props.children}.  However, in the newest version of react-router, we can
contain nested routes within the Parent component themselves instead of the
Router component. Declaring a route within a component is equivalent to
rendering a regular component in that they are both rendered as children. It
indicates where it should be rendered in the component if active. You can
declare as many routes as you need, but only one may be active at a time which
means only one will be rendered at a time. Don't take this to mean that only one
child will be rendered, the active rule only applies to routes. This seems like
a more intuitive approach, but it's a confusing one if you are used to
centralizing all the routes under Router.

Now under Router, we add two more routes associating /home and /battle to the
Home and Battle components respectively. The Home and Battle components are also
created, but they only contain placeholder data for now. We need something to render
if we are going to have routes to them.

As for the styling, we remove some ugly details and add some color. For
instance, we remove the underline decoration for the a tags. We also change the
color of the text to charcoal -- that may change in the future. And we add
active styles to links, that is, styles for when the link's path is active. When
the link's path is active, we color it crimson. We use flexbox to create the
nav, and add some spacing between the li's. There are other small details, be
sure to look under the nav bar section in the stylesheet.

app/App.js:
        - Adding a Router, a BrowserRouter to be exact, to the application.
                - The Router component contains a div holding 4
                  components -- 1 Nav and 3 Routes.

app/componets/Battle.js:
        - Contains a placeholder until the component is built

app/components/Home.js:
        - Contains a placeholder until the component is built

app/components/Nav.js:
        - The Navigation Bar is simple unordered list consisting of three
          react-router aware links (Nav Link).
                - The react-router aware links are components that add
                  functionality to <a> by wrapping functionality such as being
                  aware of the current active path, allowing active styles,
                  and deferring to the router to route the user to the link's
                  correct path.

app/index.css:
        - Adding styling to the Navigation Bar
                - Flexbox is used to create the nav bar
        - Changing the text color to charcoal and removing the underlines for
          the links.

package.json:
        - Adding 'react-router-dom' package as a dependency