History API Callback
Since our app is a single page application with our index page being the only
page served by the server, any requests made by the browser to a path other than
the index will result in a "cannot get"/404 message. This also happens when we
refresh the page at a path other than the index. Routing isn't handled by the
server in our application, instead, react-router handles our routing. To help us
to navigate to different endpoints, we defer to HTML5's history API and direct
webpack's dev-server to use it.

When an endpoint is reached, the request the browser makes is changed to a
request for the index page and allows react-router to take over the routing.
There are other options that allow us to configure how to direct requests to
certain paths.

That's sort of the gist of it.

/webpack.config.js:
        - Adding the historyApiCallback feature for webpack-dev-server