404 for Non-Existent Paths
Now that we've reconciled manually entering paths in the browser with
react-router by using the historyApiFallback API, we add a switch to catch all
paths that don't match the existing routes. Using a switch, a single route is
active at any given time. Therefore, we place a catch-all route at the end that
maps to a 404 message. If none of the routes match, we fall through to the last
catch-all route. This way we render a 404 message for all non-existent paths
requested by the user.

app/components/App.js:
        - Adding Switch to render routes when the path matches, and default to a
          404 when the path doesn't match to any existing route.

app/components/NotFound.js:
        - A simple 404 message alerting the user that the path doesn't exist

app/index.css:
        - Styling for the NotFound Component