Form Handling in the Battle Component
We envision Battle encapsulating two forms that allow users to submit their
usernames so that they can compare their GitHub stats against each other. It seems
simple enough, however, handling the state of the components is rather tricky.

First we'll focus on the form component and then we'll discuss Battle's role in
maintaining the inputs from these forms.

Normally when we have a form, we would have the user type something into the
input box(es) and hit submit. We could then use the default action of submitting
an HTTP request to the server (with a path and a HTTP method specifying what
action the server should take), or we could prevent the default action and,
using the DOM, grab the value from the input box on submit (on a submission).
Here the latter sounds apropos since our server doesn't handle routing (to a
handler using the path and HTTP method). Since we're using React, we take
advantage of react's state handling to update the username typed by the user
rather than reaching into the DOM for the input box's value. This is a
controlled way for dealing with input forms, that is, react is controlling the
state of the input forms.  Reaching into the DOM on submit is an uncontrolled
way for handling form -- react has no control over the state of the form.

In having React control the state of our forms, we create method changeUsername
to update the state of the form instance when a change occurs in the box. The
state is comprised of just the username, the input corresponding to the input
box. This method is a callback used by the input box to update the state of the
form component. Passing a callback to a child to communicate to the parent to
update its state is common pattern in React and we'll use it again later.

The form component's render method is responsible for rendering the UI, the form
UI. The render method is pretty simple, it renders a form with an input box and
button. The input box is passed the changeUsername callback to update the form's
stored username. It also replaces the value in the box with the stored username
when a re-render occurs because of the call to setState by changeUsername. The
form itself has an action on submit to call a callback that will communicate to
the parent Battle to update it's state. We discuss the implementation of
handleSubmit now.

At this point, you might ask following the question: why don't we just assign
onSubmit the handler passed by Battle? We define another method because it's not
only a matter of passing a callback, we also need to prevent the form from
performing its default action. So we create a function called handleSubmit that
defines this procedure of: 1. event.preventDefault() which prevents the on
submit event from triggering the form's default action.  2. invoking Battle's
callback to tell it to update its state We pass the method as a callback through
the form's onSubmit prop -- remembering to bind the method to the instance
because we make references to "this"; we invoke the method in a different
context, therefore, the object "this" refers will change unless we bind "this"
to the instance.

Now we discuss Battle's role in maintaining the state. Up to this point, we
built the UserInputForm without knowing what arguments, if any, the callback
method accepts. All of this depends on the state and how we update our state.
The state is comprised of 4 fields: user1 and user2, the usernames of two users;
and the user1Avatar1 and user2Avatar, the corresponding avatars for the users.
To update this state, we define handleSubmit. This will be the callback we send
to our children forms. However, we have a problem. This function is suppose to
maintain the state of two different submitted inputs, how do we know which form
submitted the input, causing the invocation of handleSubmit? We pass the forms
some id through props that they can use to identify themselves upon invoking
handleSubmit. We make handleSubmit ask for this id upon invocation, in addition
to the username being submitted.

The handleSubmit uses the id and the new username to update both the username
and avatar corresponding to the user who submitted it -- we'll generate a url to
an avatar image rather than storing an actual image. We pass this callback to
the children forms -- remembering to bind the method to an instance of Battle in
the constructor.  (It makes sense to permanently bind a method to the instance
during its construction.)

NOTE: The url's aren't generated just yet. There is a placeholder at the moment.

The render method is pretty simple. We render two form components,
passing in an id for each of them; passed with the id is handleSubmit to allow
the children to communicate to Battle to update its state with new info given to
them, and a label to display with the form.

When a username is submitted, Battle's state will have a username other than ''.
We can use this to our advantage to check if there was a submission or not. If
there was a submission, we replace the form with text saying "Submitted".
Otherwise, there was no submission and we should display a form.

Next, we'll generate urls for the avatars, style the forms, and create a proper
submission component that will replace the submitted placeholder.

app/components/Battle.js:
        - Adding a Form component called UserInputForm, which maintains is own
          state.

        - The Battle component maintains a state comprised of the usernames of
          two players, each submitted by the forms, and their corresponding
          avatars (the usernames are used to generate the urls to the user's
          avatar).
                - The urls aren't generated yet. I will have to look into the
                  GitHub API (there is a simple url that we can use to
                  access the avatar image).

        - Because the form component performs the submit action, we need some
          way for the forms, children of Battle, to communicate to the parent to
          change its state.  The pattern we use is the same as the one we
          used for implement the language bar: we define a function/method
          to update the state, bind a function/method to an instance of a
          component, and pass the function as a callback down to the children.
          The children can invoke the callback when needed to change the
          parent's state.
                - We use this pattern to replace the default submit action of
                  the form. We create handler that disables the default submit
                  action and invokes the callback passed down by Battle.

        - The forms aren't styled yet.