Battle Component Styling
I was trying to figure out how I could make the Battle component look less
boring without creating my own assets (I suck at art and SVGs). Thankfully, Font
Awesome exists so I included the CDN in the index page. I only needed it for the
GitHub icon, so it may seem like overkill to add the entire library, but I may
use more of it later on.

I wanted to use the icon to convey that we are dealing with GitHub users. Note
that since we are using icons, this kills the accessibility of the page.  I know
accessibility shouldn't be an afterthought, but this app is just a proof of
concept after all. The problem with using the icon is that it doesn't label the
form. It doesn't convey that we expect two different users; the form is just two
GitHub icons with input boxes. I experimented with adding text above it and
below it. Also, I considered keeping the word "user" in the label.  But then I
saw that the GitHub icon had enough space in the head of the silhouette to fit
text.  I figured if I could increase the size of the silhouette, then I'll have
enough space to fit a number within the silhouette's head. The number will
convey to which user the form belongs. So I blew up the icon, by increasing the
font size (the icon is considered a font character). The problem now was
overlaying the number over the SVG.

To overlay text, specifically a number over an element, I had to absolutely
position the text to get it out of the flow and keep it from stacking under the
icon. To contain the absolutely positioned text, the icon element is set to
relative positioning; this way, text is positioned relative to icon and not some
other nearby positioned ancestor. Next, I had to center the text. I had to do a
little math to get this right: take the width of the text's box, subtract that
from the width of the entire box to get the space left over. We then divide the
leftover space into two equal parts -- the division results in the space, in px,
that should be on either side. This is the value that right (or left) should
take. I used trial and error (using the browsers dev tools) to find the closest
percentage that results in the value we calculated. Why did I use percentages
instead of px? If we increase the font-size in the browser, the SVGs expand. As
the SVGs expand, the amount of px necessary to center the text changes.
Therefore, we can't rely on static units. If the SVG was kept static, then we
could use px to position the text.

The last problem that I had to solve was getting the appropriate text to
display. We couldn't use ::after to help insert the text over the icon because
the content property is fixed. So, instead, we used a ternary operator to return
the appropriate text element as a child of the icon element. This is the reason
why we set the icon element's positioning to relative, to capture the absolutely
positioned child.

The rest of the styling is straight forward. One thing we have to note is that
.fa overrides the font settings of less specific rules. We have to make sure to
override fa's settings when appropriate.

I also use relative font sizes for the fonts. I don't use a relative padding for
the input box because the input box is static. However, I do set the margins of
the input box to be relative. I want the spacing to be consistent with the size
of the text, though, now that I think about it, I don't think it's necessary. (I
should probably play around with this and see what happens.)

I'm still a little unhappy with the text. Perhaps I'll experiment a little more
with different fonts, font-sizes, and font-weights.

app/components/Battle.js
        - Replacing the text labels ("User 1" and "User 2") with SVG icon labels
                - The icon's used are from the Font Awesome font/icon library
                - We add a ternary operator to display the appropriate text for
                  the user's form. The text is going to be absolutely positioned
                  over the SVG so that it is visible and centered.

index.css:
        - Styling the Battle Component and the UserImportComponent

index.html:
        - Adding the script that has access to our Font Awesome CDN (we include
          Font Awesome as part of our page)