Styling the UserInputComponent
I accidentally forgot include the values necessary to center the numbers within
the icon. The previous values were a result of eyeballing the text to make sure
it was centered. The values here, on the other hand, were derived using simple
arithmetic. Also, I found a good weight and size for the font that makes look
like it's a part of the icon.

index.css:
        - Changing the font and font-weight of the label text (the numbers)

        - Changing the values used to center the absolutely positioned label
          text (the numbers)