Implementing Previews and Setting Up Results Comp.
Remember that "Submitted" placeholder we display when a user submits a username?
It's being replaced with a preview component that shows the submitted user's
avatar and his username. The component is pretty simple, it composed of an image
and a header wrapped in a div. The image source and the username is passed down
from the Battle component's state into the PlayerPreview component as props.

The image url is generated using the state's username. I hadn't previously
implemented this because I thought a fetch using the GithHub API would be
necessary. Luckily, the GitHub has a url for user avatars up to 400px, I
believe. You can append the '?size=' query string with a lower size if you like.

We also create a Results Component that contains a placeholder as prelude to
implementing the next stage of the Battle section. We want to be able to pass
the submitted usernames over to the Results component using a url with a query
string. The Results Component during after being mounted on the DOM can parse
the query string and fetch the stats using the GitHub API. Then we'll re-render
the component with the fetch stats. That's what we want to accomplish next.

Additionally, we render a battle button that will confirm the battle and
initiate the routing to /battle/results. The button will be responsible for
passing over the submitted usernames by providing a query string to the Link
component so that it can be appended to the path.

app/components/App.js:
        - Adding a new route for the path '/battle/results'; the component
          associated with this router is Results

        - Modifying Battle route to be an exact route. the root of the results
          path contains /battle. Any path containing /battle will be matched to
          the /battle Route component

app/components/Battle.js:
        - Adding the PlayerPreview Component

        - Adding a button that routes the client to '/battle/results'
                - The button in the future will provide a query string with the
                  submitted user names in the state.

app/components/Results.js:
        -  It contains placeholder text. This will be replaced with the
           submitted users' stats.

app/index.css:
        - Styling modifications
                - Adding text-align to .button